Renamed tests and folders to match stories
