+++ tests/Sidebar/103-SectionsList.test.tsx
